APIs/bbc.com/1.0.0/python-aiohttp-client/tests/test_raw_controller.py
# coding: utf-8

import pytest
import json
from aiohttp import web

from openapi_server.models.nitro import Nitro


pytestmark = pytest.mark.asyncio

async def test_get_raw_ancestors(client):
    """Test case for get_raw_ancestors

    Get raw ancestors
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/episodes/{pid}/ancestors'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_brand(client):
    """Test case for get_raw_brand

    Get raw brand
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/brands/{pid}'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_brand_franchises(client):
    """Test case for get_raw_brand_franchises

    Get raw brand franchise
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/brands/{pid}/franchises'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_episode(client):
    """Test case for get_raw_episode

    Get raw episode
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/episodes/{pid}'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_formats(client):
    """Test case for get_raw_formats

    Get raw formats
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/episodes/{pid}/formats'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_genre_groups(client):
    """Test case for get_raw_genre_groups

    Get raw genre groups
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/episodes/{pid}/genre_groups'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_image(client):
    """Test case for get_raw_image

    Get raw image
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/images/{pid}'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_masterbrand(client):
    """Test case for get_raw_masterbrand

    Get raw masterbrand
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/master_brands/{mbid}'.format(mbid='mbid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')


pytestmark = pytest.mark.asyncio

async def test_get_raw_promotion(client):
    """Test case for get_raw_promotion

    Get raw promotion
    """
    headers = { 
        'Accept': 'application/json',
        'api_key': 'special-key',
    }
    response = await client.request(
        method='GET',
        path='/v1/promotions/{pid}'.format(pid='pid_example'),
        headers=headers,
        )
    assert response.status == 200, 'Response body is : ' + (await response.read()).decode('utf-8')

